Moved main process code to new folder
Also improved dialog prompts for opening/downloading siad
